Document required packages after conda update
The number of dependencies has gone down now that jupyterlab has been updated. ipywidgets may not be the minimum package requirement (perhaps one of its dependencies is what allows the new environment to propagate to jupyterlab). However I don't really have time to check this and it works ok at the moment.